requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-Cors==3.0.10
python-dotenv==0.21.0
pandas==1.5.3
openpyxl==3.0.10
